Begin separating socket routes (lots of websocket issues to fix)
